remove api, sdk, and ui packages from dependabot
sdk and api are removed because those two are libraries, and updating libraries to the latest
dependency versions only serves to make the library harder to use. A
library which uses all the latest versions requires every application
which uses it to update all its dependencies in a single batch, which is
not great.
It is applications that need the latest version of dependencies.

Remove UI at the request as jc, since javascript libraries are updated so frequently it ends up
being too much noise.

Also update the number of PRs for gomod to 10, so we have a few more to
work with.