fix: customize eai publicAssetURL in production
The `lib/startup`-addon is used to create a custom
index.html page to boot up the app. Because this is
a custom approach and we aren't relying on setting
rootURL to a different value when running the app
in `production` we need to set `publicAssetURL`
in the configuration we pass to `ember-auto-import`.

Reference:
https://github.com/ef4/ember-auto-import/blob/9a2887f1d09c1df2a41d7e6800820ba23183365e/docs/upgrade-guide-2.0.md#quick-summary